+++ tests/test_blob_migrate.py
@@ -7,6 +7,7 @@ from pathlib import Path
 from . import S3_CONN_INFO
 from . import CONN_INFO
 from datajoint.migrate import _migrate_dj011_blob
+dj.config['enable_python_native_blobs'] = True
 
 
 class TestBlobMigrate: