+++ tests/test_jobs.py
@@ -1,51 +1,51 @@
-from decimal import Decimal
 from nose.tools import assert_true, assert_false, assert_equals
 from . import schema
 from datajoint.jobs import ERROR_MESSAGE_LENGTH, TRUNCATION_APPENDIX
 import random
 import string
+import datajoint as dj
+dj.config['enable_python_native_blobs'] = True
 
 subjects = schema.Subject()
 
 
 def test_reserve_job():
-    # clean jobs table
-    schema.schema.jobs.delete()
 
+    schema.schema.jobs.delete()
     assert_true(subjects)
     table_name = 'fake_table'
+
     # reserve jobs
     for key in subjects.fetch('KEY'):
-        assert_true(schema.schema.jobs.reserve(table_name, key),
-                    'failed to reserve a job')
+        assert_true(schema.schema.jobs.reserve(table_name, key), 'failed to reserve a job')
+
     # refuse jobs
     for key in subjects.fetch('KEY'):
-        assert_false(schema.schema.jobs.reserve(table_name, key),
-                     'failed to respect reservation')
+        assert_false(schema.schema.jobs.reserve(table_name, key), 'failed to respect reservation')
+
     # complete jobs
     for key in subjects.fetch('KEY'):
         schema.schema.jobs.complete(table_name, key)
-    assert_false(schema.schema.jobs,
-                 'failed to free jobs')
+    assert_false(schema.schema.jobs, 'failed to free jobs')
+
     # reserve jobs again
     for key in subjects.fetch('KEY'):
-        assert_true(schema.schema.jobs.reserve(table_name, key),
-                    'failed to reserve new jobs')
+        assert_true(schema.schema.jobs.reserve(table_name, key), 'failed to reserve new jobs')
+
     # finish with error
     for key in subjects.fetch('KEY'):
-        schema.schema.jobs.error(table_name, key,
-                                 "error message")
+        schema.schema.jobs.error(table_name, key, "error message")
+
     # refuse jobs with errors
     for key in subjects.fetch('KEY'):
-        assert_false(schema.schema.jobs.reserve(table_name, key),
-                     'failed to ignore error jobs')
+        assert_false(schema.schema.jobs.reserve(table_name, key), 'failed to ignore error jobs')
+
     # clear error jobs
     (schema.schema.jobs & dict(status="error")).delete()
-    assert_false(schema.schema.jobs,
-                 'failed to clear error jobs')
+    assert_false(schema.schema.jobs, 'failed to clear error jobs')
+
 
 def test_restrictions():
-    # clear out jobs table
     jobs = schema.schema.jobs
     jobs.delete()
     jobs.reserve('a', {'key': 'a1'})
@@ -54,11 +54,11 @@ def test_restrictions():
     jobs.error('a', {'key': 'a2'}, 'error')
     jobs.error('b', {'key': 'b1'}, 'error')
 
-    assert_true(len(jobs & 'table_name = "a"') == 2, 'There should be two entries for table a')
-    assert_true(len(jobs & 'status = "error"') == 2, 'There should be two entries with error status')
-    assert_true(len(jobs & 'table_name = "a"' & 'status = "error"') == 1,
-                'There should be only one entries with error status in table a')
+    assert_true(len(jobs & {'table_name': "a"}) == 2)
+    assert_true(len(jobs & {'status': "error"}) == 2)
+    assert_true(len(jobs & {'table_name': "a", 'status': "error"}) == 1)
     jobs.delete()
+
 
 def test_sigint():
     # clear out job table
@@ -73,6 +73,7 @@ def test_sigint():
     assert_equals(error_message, 'KeyboardInterrupt')
     schema.schema.jobs.delete()
 
+
 def test_sigterm():
     # clear out job table
     schema.schema.jobs.delete()
@@ -86,6 +87,15 @@ def test_sigterm():
     assert_equals(error_message, 'SystemExit: SIGTERM received')
     schema.schema.jobs.delete()
 
+
+def test_suppress_dj_errors():
+    """ test_suppress_dj_errors: dj errors suppressable w/o native py blobs """
+    schema.schema.jobs.delete()
+    with dj.config(enable_python_native_blobs=False):
+        schema.ErrorClass.populate(reserve_jobs=True, suppress_errors=True)
+    assert_true(len(schema.DjExceptionName()) == len(schema.schema.jobs) > 0)
+
+
 def test_long_error_message():
     # clear out jobs table
     schema.schema.jobs.delete()
@@ -96,7 +106,7 @@ def test_long_error_message():
     assert_true(subjects)
     table_name = 'fake_table'
 
-    key = list(subjects.fetch('KEY'))[0]
+    key = subjects.fetch('KEY')[0]
 
     # test long error message
     schema.schema.jobs.reserve(table_name, key)